+++ src/titiler/core/titiler/core/factory.py
@@ -1667,6 +1667,15 @@ class AlgorithmFactory:
             """Algorithm Metadata"""
             props = algorithm.model_json_schema()["properties"]
 
+            # title and description
+            info = {
+                k: v["default"]
+                for k, v in props.items()
+                if k == "title" or k == "description"
+            }
+            title = info.get("title", None)
+            description = info.get("description", None)
+
             # Inputs Metadata
             ins = {
                 k.replace("input_", ""): v["default"]
@@ -1685,9 +1694,18 @@ class AlgorithmFactory:
             params = {
                 k: v
                 for k, v in props.items()
-                if not k.startswith("input_") and not k.startswith("output_")
+                if not k.startswith("input_")
+                and not k.startswith("output_")
+                and k != "title"
+                and k != "description"
             }
-            return AlgorithmMetadata(inputs=ins, outputs=outs, parameters=params)
+            return AlgorithmMetadata(
+                title=title,
+                description=description,
+                inputs=ins,
+                outputs=outs,
+                parameters=params,
+            )
 
         @self.router.get(
             "/algorithms",